Do not notify the SSE server on comment added to a ticket via git
We should notify the SSE server even on update of the ticket via git
push to the ticket repo (the only case where notify=False basically),
but this causes problem with some of our markdown extension so until we
figure this out, we won't do live-refresh.

Eventually, we could do the markdown conversion in its own service and
store the results in the DB where we would then retrieve it.