When creating the pull-request, save the commit_start and commit_stop
If we were able to compute a list of commits differing between the two
repo there is no reason not to save that info in the DB from the start.

In addition this will be most useful since it will be announced on the
fedmsg message about this new PR and consumer will be able to access and
use that info.