Fixing showing tags even when some of them are not formatted as expected
We encountered in https://pagure.io/pagure/issue/1991 an issue where some
tags objects were not formatted as libgit2 expects them to be.
In this case, there were no email attached to the person doing the tag,
the tagger, and libgit2 would raise an ValueError exception on this.
With this commit, since we cannot access the object, we just catch this
exception and move on.
It's a pity since we will not be able to show these tags, but at least
we will be able to show the rest of them.

Fixes https://pagure.io/pagure/issue/1991