Support irc:// links in our markdown and adjust the regex
In this commit we are improving our markdown processor in different
aspects:
- Replace the regex to match links with the one coming from the markdown
project itself, adjusted for our needs.
- Add support to link on irc:// and ircs:// in markdown, this is however
only partly working because the rendered markdown is passed onto
python-bleach which cleans the ircs links since ircs isn't in the  list
of its supported protocols. More recent version of python-bleach (from
1.5 basically) would allow us to expand the list of supported protocols
and add ircs to it, but the version we are currently using (1.4.3) does
not offer that option.

Fixes https://pagure.io/pagure/issue/2085