Fix unicode in password and add test
After a long discussion with ``@jcline`` and as he pointed out that using
``to_unincode`` is a really bad option for storing password since after
unicoding some non-ascii value it starts giving the same character. This could
be really harmful for security reasons.

The tests are provided to check the same that things don't break on entering
any non-ascii characters.

Signed-off-by: Farhaan Bukhsh <farhaan.bukhsh@gmail.com>