Rename "request pull" to pull request
People are used to the "pull request" name, use it instead,
"request pull" is confusing and brought major language issues
in most cases.

As this is more or less a text fix, also fixed a typo on
issues page (s/on/of in open/closed issues bar) and reworded it
to sound a bit better.

Merges: #2144