Add a spinner indicating when we are retrieving the list of branches differing
Basically, on the drop-down menu present on the page listing all the
pull-request, we do an ajax call to find out what are the different
branches for which we can create a pull-request. Since may take a little
time depending on the number of branches and the load on the server, so
this commit adds a small spinner indicating there is something happening
and the user should wait a little.

Fixes https://pagure.io/pagure/issue/2166