Fix tests and migrated to ``_getrepo``
The ``issue`` api changed its architecture to use a different way of getting
the repo. The refactoring is amazing since now only one function is required to
make a get repo call.

This commit also fixes few test cases and modified the test json for the same.
The alembic revision file is updated with the new revision.

Signed-off-by: Farhaan Bukhsh <farhaan.bukhsh@gmail.com>