Link to the doc when the documentation is activated but has no content
New user of pagure may not know how the documentation works in pagure.
So they will see the settings page, activate the documentation and if
they do not upload anything to the documentation git repo, nothing will
happen and they may wonder why.
So in addition to the error message that nothing was found in the git
repo, this commit adds a link to the relevant section of pagure's
documentation.

Fixes https://pagure.io/pagure/issue/1562

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>