Fix pushing merge commits to the original repo
For some reason pygit2 wasn't properly creating the local reference to
be pushed to the main repo. With this commit we enforce that all merge
commits use a local reference that is then pushed to the target branch.

Fixes https://pagure.io/pagure/issue/1190
Fixes https://pagure.io/pagure/issue/2119

Thanks to @puiterwijk for the team work on tackling this one

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>