commit instead of flush on issue assign update
Like it does earlier in the file, when resetting the assignee. If we
don't commit the transaction, SA won't let us access the updated
attributes, which is why we were (correctly) getting None in the reset
case (since it calls commit), but we were getting the old assignee in
the reassign case.

This fixes #1896.

Signed-off-by: Ricky Elrod <ricky@elrod.me>