Introduce the external committer feature
This feature allows to grant commit ACL to user via groups that are
specified by the authentication system upon login.
This allows by-passing the group mechanism of pagure itself, thus allowing
to grant commit access without creating the corresponding group and
without notifying every member of that group for changes to project they
have access to.

This feature allows to grant access to all the projects of the instance,
only some (using `restrict`) or to all but some (using `exclude`).

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>