Add ways to customize the gitolite configuration file with snippets
This commit adds two configuration key, GITOLITE_PRE_CONFIG and
GITOLITE_POST_CONFIG allowing to point to files that will be included
at the top or at the end of the gitolite configuration file.
This allows customizing the gitolite configuration file with elements
or information that are outside of pagure's control.
This can be used in combination with the ``EXTERNAL_COMMITTER`` to give
people of certain groups commit access to the git repositories.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>