Fix diffing the branch of a project against the target branch
We used to check the local branches against the local default branch (ie
most often master) but this failed in a few situation:
a) when the master of the fork was lagging behind the one of the main
repo (in which case it would offer a PR for things that had already
been included)
b) when the changes in the fork were actually in the master branch
(since it skipped master to master comparison).

With this commit we are now comparing the local branches to the default
branch of the main repo, still skipping comparing default to default if
the repo is not a fork.

Fixes https://pagure.io/pagure/issue/2354

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>