When loading comment from JSON rely on username/comment rather than comment id
Instead of relying on comment id which may vary from instance to instance
depending on what is in the database, use the combo username/comment to
find out if a comment was already made to the ticket (so that if not, it
gets created).
We are short-coming the case where the same user submits twice the exact
same comment, so be it.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>