Fix the api_view_user_activity_stats to return the expected data
In flask 0.11 quite some changes were made to flask.jsonify() changing
its behavior and in this case breaking the calendar heatmap widget.
This commit fixes this.

While at it, highlight in the calendar heatmap today just to make it
prettier.

Let's not speak about the change to the tests here, not now, not ever...

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>