Bail on merge a PR that is already closed
It happened at least once that two persons requested to merge a PR at
the same time, leading to two comments being added to the PR about it
being merge (by each individual) while the task for the second person
actually failed (because there was nothing to change).

With this change, we're waiting for the last moment and checking if the
PR has been merged/closed and if so, we bail.

Fixes https://pagure.io/pagure/issue/2686