Make the pagure hook act as the person doing the push
This was a small annoyance we had, if you were amending a commit from
someone else, or just rewording the commit message (for example to add
the ``Merges...`` line) and pushing to the main repo, pagure would show
the action has being done by the author of the commit rather than the
person who pushed that commit.

This commit fixes this by relying on the environment variable set by
gitolite if it is present, else going back to the author of the commit
if we cannot know who pushed.

Fixes https://pagure.io/pagure/issue/2597

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>