Return the flag UID when adding or updating a flag on a PR not in fedmsg
The UID of a flag is what allows to update it, so it shouldn't be part
of the fedmsg message we sent.
However, if the user is not specifying the UID, then they have no way to
know it unless we return it, so when adding or updating a flag, return
the JSON representation of that flag object.

Adjust the unit-tests accordingly.

Fixes https://pagure.io/pagure/issue/1588

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>