Add a status field to flags
This status is used in addition to the percentage making it optional.
For flags on commit the status is mandatory, on pull-requests it is
optional and will be inferred from the percentage specified (or not).

This keeps the backward compatibility on the API endpoint to flag PR
while enforcing the new behavior for the new API endpoint for commits.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>