Use the updated timestamp in the pull-request list
Using the updated_on timestamp instead of the last_updated we are
able to give the users a better idea of when exactly the PR got
updated.
The last_updated field is updated when any other PR is merged since
then the merge_status cached value is reset.

Fixes https://pagure.io/pagure/issue/2725

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>