Do not make optional dependencies required
This content of the requirements.txt file is processed in the setup.py
and added/declared as being what pagure requires to work.
But, if you deploy pagure, say with openid login, you do not need
flask-oidc, but if you keep it in the list, it's added to the .egg-info/
requires file as being required and when something queries pkg_resources
it will complain that flask-oidc is required and missing and make the
app crash.

So let's keep optional packages not required :)

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>