Fix redirecting the user when they remove themselves from a project
We used to redirect the user to the settings page but if the user remove
themselves from the project, they can no longer access this page and they
end up with a nice 403 page.
With this commit, we check if they are removing themselves and if they
do we simply redirect them to the front page of the project.

Fixes https://pagure.io/pagure/issue/2937

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>