docs: pr: use origin remote name instead of upstream
In 99% of cases, the remote it called 'origin'. It seems silly to have a
differently named remote in example docs, especially when it contains a
hint to copy over a particular line into git config files, that requires
the users to change the line over and over for each time they use it.

Also improve formatting and some whitespace.

Signed-off-by: Kamil Páral <kparal@redhat.com>

Merges https://pagure.io/pagure/pull-request/2950