Fix editing and deleting comments added by the EV server to PRs
Basically, we had the project url hard-coded in the html instead of have
it automatically generated from the content passed on by the EV server
so editing a comment added by the EV server was only ever going to work
on a one project.

In addition, the form and csrf token were missing to allow the delete
button to work.
Both form and csrf token are now present on the html page, thus allowing
to right away delete a generic comment that was added by the EV server.

Fixes https://pagure.io/pagure/issue/2958

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>