Deduplicate list of contributors
It is possible the same person commits into a repo with multiple
identities. Pagure already can be told about multiple e-mails for a
single person. This patch checks if the address in the commit is known
and possibly replaces with the default e-mail.

If the e-mail is known, the name is also replaced with full name of the
user.

Records for contributors not known to Pagure are presented as before.

The page showing commits for a single contributor already displays all
of them no matter what identity was used.