Return a 418 error if the task did not return anything itself
We need this endpoint to always return something but there are tasks that
are not returning an endpoint, so we can't redirect the users to anywhere
as we do for the other tasks. So instead, in these case, we'll raise a
418 error.
This was faced while working on tests and did not happen in real life.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>