Fix the user's requests page
We used to include the projects in which an user had opened at one point
a pull-request in, this lead to a lot of false positive.
Now we're being strickter about when retrieving the list of projects
and adding the pull-requests that the used opened in a second step.

So the logic is something like:
- Get all the projects the user is involved in (direct access or via a
  group)
- Get all the PRs in these projects + all the PRs opened by that user
- Group them all into one final query that is then ordered/filtered as
  desired

Fixes https://pagure.io/pagure/issue/2930

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>