Fix logging commits in docker-container dev env
This fixes up several problems with logging commits in the
docker dev env. We did not have an equivalent of pagure_logcom
running at all, first of all. There were also several issues
preventing the default commit hook from running properly. The
file is not executable, which meant the hook wouldn't be run
at all (@cverna noticed this). PYTHONPATH not being set as an
env var in the web container meant the hook couldn't find the
pagure module. And PAGURE_CONFIG not being set as an env var in
the web container meant the hook couldn't find the correct
pagure config file for the dev environment.

With all these things fixed, if you clone a repo from the dev
instance *inside* the web container (it currently only works
that way) and then push a commit to it (using an email address
associated with a user that exists), the hook runs, and the
commit is properly logged to the database, and shows up in the
user's activity log.

Signed-off-by: Adam Williamson <awilliam@redhat.com>