doc: explain how to fetch a pull request locally
When working with pull requests it is easy to use `git fetch` to get a
single pull request locally.  Explain how to do this before detailing
how to edit the `.git/config` file for fetching all pull requests.

Update the tip on the pull-request page to suggest `git fetch` to a
local branch followed by a `git checkout` of the branch.