Add comment reactions
A user can select from a predefined list of reactions for a comment.
There can be multiple reactions from the same person on a single
comment, but only one of each type.

The available reactions are listed in the config file as a human
readable name (which will ultimately be stored in database) and a
emojione class used for displaying.

This works for both issues and pull requests.

The reactions are visible in the API, but can only be posted via the web
UI.

Fixes: https://pagure.io/pagure/issue/812