Add a way to list all API tokens and don't restrict the info command
By default `pagure-admin admin-token list` will only return API token
that have at least one of the ACLs defined in: ADMIN_API_ACLS.
This action now supports an `--all` options making it return all API
tokens period.

In addition, the `pagure-admin admin-token info` had the same restriction
but for the info action the API token must be already known, so there is
no point in restricting accessing this info based on the ACLs of the token.
So just return the info of that token regardless of the ACLs associated
with it.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>